Form add

+++ server/models/substance.js
@@ -2,7 +2,7 @@
 module.exports = (sequelize, DataTypes) => {
   var Substance = sequelize.define('Substance', {
     name: DataTypes.STRING,
-    points: DataTypes.INTEGER
+    points: DataTypes.FLOAT
   }, {tableName: "substances"});
   Substance.associate = function(models) {
     // associations can be defined here
